refactor: step1 테스트 패키지 분리
