+++ src/test/java/step1/SetTest.java
@@ -1,4 +1,4 @@
-package lotto;
+package step1;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
